change default int type to i64
numpy defaults to 64-bit integers on most architectures. Also improves error messages when chaining fails to include the atom type.

+++ python/test/test_metric_cast.py
@@ -14,33 +14,33 @@ enable_features("contrib")
 
 def test_ordering():
     data = [1, 2, 3]
-    ord_trans = make_ordered_random("i32")
+    ord_trans = make_ordered_random("i64")
     assert len(ord_trans(data)) == 3
 
-    ident_trans = ord_trans >> make_unordered("i32")
+    ident_trans = ord_trans >> make_unordered("i64")
     assert len(ident_trans(data)) == 3
 
 def test_sized_ordering():
     data = [1, 2, 3]
-    ord_trans = make_sized_ordered_random(3, "i32")
+    ord_trans = make_sized_ordered_random(3, "i64")
     assert len(ord_trans(data)) == 3
 
-    ident_trans = ord_trans >> make_sized_unordered(3, "i32")
+    ident_trans = ord_trans >> make_sized_unordered(3, "i64")
     assert len(ident_trans(data)) == 3
 
 def test_sized_bounded_ordering():
     data = [1, 2, 3]
-    ord_trans = make_sized_bounded_ordered_random(3, (3453, 23245), "i32")
+    ord_trans = make_sized_bounded_ordered_random(3, (3453, 23245), "i64")
     assert len(ord_trans(data)) == 3
 
-    ident_trans = ord_trans >> make_sized_bounded_unordered(3, (3453, 23245), "i32")
+    ident_trans = ord_trans >> make_sized_bounded_unordered(3, (3453, 23245), "i64")
     assert len(ident_trans(data)) == 3
 
 def test_bounded():
     data = [1, 2, 3]
 
-    bdd_trans = make_metric_bounded(3, "i32")
+    bdd_trans = make_metric_bounded(3, "i64")
     assert len(bdd_trans(data)) == 3
 
-    ident_trans = bdd_trans >> make_metric_unbounded(3, "i32")
+    ident_trans = bdd_trans >> make_metric_unbounded(3, "i64")
     assert len(ident_trans(data)) == 3
